Fix the notebook deploy script.
Fixed kwargs typo (an obvious thus difficult one). Switched to regex instead of `startswith()` to circumvent the script to fail on blank chars due to code alignment (a nasty one). Had to deal with newline `\n` handling as if a line is skipped no extra \n should be added. To circumvent Franklin's Literate preprocess, I also process the _web scripts that are deployed for the website with rm_hint/sol.